refactor: 쿠폰 사용 가능 여부 확인 커스텀 훅으로 분리
